marge to 85x

+++ path_8_5x/sources/creature.h
@@ -234,8 +234,8 @@ class Creature : public AutoId, virtual public Thing
 
 		int32_t getWalkDelay(Direction dir) const;
 		int32_t getWalkDelay() const;
-		int32_t getStepDuration(Direction dir) const;
-		int32_t getStepDuration() const;
+		int64_t getStepDuration(Direction dir) const;
+		int64_t getStepDuration() const;
 
 		int64_t getEventStepTicks(bool onlyDelay = false) const;
 		int64_t getTimeSinceLastMove() const;
